+++ simconnect-sdk/src/domain/mod.rs
@@ -1,6 +1,6 @@
 mod condition;
 mod data_type;
-mod event;
+mod events;
 mod facilities;
 mod notification;
 mod notification_group;
@@ -8,7 +8,7 @@ mod period;
 
 pub use condition::*;
 pub use data_type::*;
-pub use event::*;
+pub use events::*;
 pub use facilities::*;
 pub use notification::*;
 pub use notification_group::*;